requirements.txt
google-generativeai>=0.7.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
rank_bm25>=0.2.2
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
docling>=2.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
numpy>=1.26.0
